WIP: refactoring parser, maps manipulation (needs further refactoring to postprocessor input class)
